requirements.txt
azure-cognitiveservices-vision-customvision==0.4.0
coremltools==2.0
mPyPl==0.0.3.1
Pillow==5.1.0
pipe==1.5.0
tensorflow-gpu==1.10.1
tqdm==4.19.5
pycocotools==2.0